Completed Redux refactor
Sass files are now structured according src/comp. Directory changes to entire frontend. Revise all code
