+++ client/src/reducers/index.js
@@ -1,11 +1,12 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 import authReducer from './auth_reducer';
+import invoiceReducer from './invoice_reducer';
 
 const rootReducer = combineReducers({
   auth: authReducer,
+  invoice: invoiceReducer,
   form: formReducer,
 });
-
 
 export default rootReducer;
