+++ client/src/sass/components/_all.scss
@@ -1,11 +1,10 @@
 @import '_landing';
-@import '_invoices';
+@import 'invoices/allinvoices';
+@import 'auth/_signin';
+@import 'auth/_signup';
 @import '_sidebar';
 @import '_reminder';
-@import '_signup';
 @import '_settings';
 @import '_billing';
-@import '_signin';
-@import '_delete';
-@import '_updateinvoice';
-@import '_datainvoice';
+
+
