+++ client/src/sass/components/invoices/_allinvoices.scss
@@ -1,0 +1,5 @@
+@import '_updateinvoice';
+@import '_datainvoice';
+@import '_delete';
+@import '_invoices';
+@import '_viewinvoice';
