+++ client/src/sass/components/invoices/_delete.scss
