+++ client/src/sass/components/invoices/_viewinvoice.scss
@@ -1,0 +1,84 @@
+
+.view-invoice {
+  display: flex;
+  background-color: #ffffff;
+  /* offset-x | offset-y | blur-radius | spread-radius | color */
+  box-shadow: 0px 7px 10px 0px rgba(0, 0, 0, 0.459);
+  width: 120rem;
+  height: 100rem;
+  margin: 0 auto;
+  position: relative;
+
+  &-main {
+    width: inherit;
+  }
+
+  &-navigation {
+    display: flex;
+    justify-content: flex-start;
+    // align-items: center;
+    background-color: #2E2E2E;
+    height: 7rem;
+
+    &_delete {
+      cursor: pointer;
+      color: white;
+      padding: 1rem;
+      margin-left: 3rem;
+      &:hover {
+        color:$color-primary;
+      }
+      
+    }
+    &_update {
+      cursor: pointer;
+      color: rgb(255, 255, 255);
+      padding: 1rem;
+      text-align: center;
+      &:hover {
+        color:$color-primary;
+      }
+    }
+  }
+  &-details {
+    display: flex;
+    flex-direction: column;
+    margin: 0 auto;
+    width: 30rem;
+    // border: 0.5px solid rgb(218, 218, 218);
+    margin: 0 auto;
+    &-total {
+      font-size: 2rem;
+      color: white;
+      background-color: $color-error;
+    }
+    &-flex {
+      display: flex;
+      // justify-content: center;
+      align-items: center;
+
+      & p:first-of-type {
+        font-size: 2rem;
+        font-weight: 700;
+        width: 15rem;
+      }
+
+      & p:nth-child(2) {
+        // font-size: 1.5rem;
+        margin-bottom: 1.5rem;
+        // margin-left: 
+      }
+    }
+  }
+
+  &-box {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin: 0 auto;
+  }
+  &-editDelete {
+    display: flex;
+  }
+}
